refactor: Pawn 클래스 리팩토링
- `Pawn` 클래스는 일반적인 이름인 `Piece` 클래스로 변경
- `Piece` 클래스는 이름을 갖는다
